% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_corr.R
\name{gee_corr}
\alias{gee_corr}
\title{Calculate the inverse of the correlation matrix}
\usage{
gee_corr(cluster2, dataframe_indepence)
}
\arguments{
\item{cluster2}{ensure that the first three columns are the ID, deletion status, and expiration time of the cluster failure time, followed by the covariates.}

\item{dataframe_indepence}{In an independent situation, the cumulative hazard function and parameter beta}
}
\value{
inverse of the correlation matrix
}
\description{
Calculate the inverse of the correlation matrix
}
\examples{
gee_corr(cluster2,dataframe_indepence)
}
