% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geemat.R
\name{geemat}
\alias{geemat}
\title{Title Calculate the  correlation matrix for clustered failure time}
\usage{
geemat(formula, data, beta_Lambda)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}. The \code{response} is a \code{Surv} object with right censoring.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula}.}

\item{beta_Lambda}{A dataframe contains regression parameters and cumulative risk functions when using an independent working matrix}
}
\value{
correlation matrix for clustered failure time
}
\description{
Title Calculate the  correlation matrix for clustered failure time
}
