% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgeecorr.R
\name{sgeecorr}
\alias{sgeecorr}
\title{Title Calculate the parameters of the clustered failure data}
\usage{
sgeecorr(
  formula,
  data,
  corrstr = "independence",
  corr_matix = NULL,
  eps = 1e-06,
  itermax = 100
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}. The \code{response} is a \code{Surv} object with right censoring.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula}.}

\item{corrstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{def}.}

\item{corr_matix}{When corrstr is \code{independence}, it does not need to be specified. When \code{corrstr} is \code{def}, it needs to be specified.When corrstr is independence, it does not need to be specified. When corrstr is def, it needs to be specified.}

\item{eps}{tolerance for convergence. The default is \code{eps = 1e-6}}

\item{itermax}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{itermax} iterations and
the estimates will be based on the last iteration. The default \code{itermax = 100}.}
}
\value{
the coefficient of the covariate, robust variance
}
\description{
Title Calculate the parameters of the clustered failure data
}
