% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgeeind.R
\name{sgeeind}
\alias{sgeeind}
\title{Calculate the parameters of the correlated failure data}
\usage{
sgeeind(cluster2, corr_inv, return_ind = TRUE)
}
\arguments{
\item{cluster2}{ensure that the first three columns are the ID, deletion status, and expiration time of the cluster failure time, followed by the covariates.}

\item{corr_inv}{Inverse of the corresponding working matrix}

\item{return_ind}{Return the cumulative hazard function in an independent situation.}
}
\value{
the coefficient of the covariate, robust variance
}
\description{
Calculate the parameters of the correlated failure data
in the case of the inverse of the independent working matrix or the inverse of the given working matrix.
}
\examples{
sgeeind(data,corr_inv,TRUE)
}
